Scripts/Layflat/requirements.txt
reportlab
pillow
PyPDF2
pymupdf